Fixed requested changes and also renamed Events.h to PhysicsEvents.h cause you right.
